+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Change categories:
 2. requester: Expose pldm_close() in header
 3. bios_table: pldm_bios_table_string_entry_encode_check(): Handle overflow
 4. bios_table: pldm_bios_table_iter_create(): Return NULL on failed alloc
+5. bios_table: pldm_bios_table_iter_next(): Invalid entry halts iteration
 
 ### Deprecated